split command_contract by domain (#4307)
Break src/command_contract.rs into focused submodules without changing the
public API or wire shapes:

- command_contract/output.rs owns response-mode, output-file, JSON-family,
  descriptor, response-plan types and the Commands::descriptor /
  response_plan / response_mode / output_file_mode /
  consumes_output_file_as_command_arg implementations.
- command_contract/lab.rs (renamed from lab_contract.rs) keeps the Lab
  portability contracts and now also owns Commands::supports_lab_runner,
  lab_runner_unsupported_reason, and lab_offload_mutation_flag plus the
  Lab-focused tests.
- command_contract/public_variants.rs holds PublicOutputVariantContract
  and the PUBLIC_OUTPUT_VARIANT_CONTRACTS table with its discriminator/
  fixture invariant test.
- command_contract.rs becomes a thin shell that re-exports the public
  surface (LabCommandContract, CommandDescriptor, CommandResponsePlan,
  PUBLIC_OUTPUT_VARIANT_CONTRACTS, etc.) so downstream call sites keep
  working unchanged.
